NOTE: UE4 BUILD VERSION: 8.3.2 &64 TODO - REMOVE GLIDE HACK AS IT IS NOW USELESS TODO - Apply glide on button trigger ('X')
Animation blending is working for the King and Boar - not the Sheep (haven't got the animations yet)
Animations/Physics on Boar need work
King is currently in ALPHA build. It moves around and feels quite nice!

Sheep needs Animations/Physics

Particles are needed (tomorrow)

Camera snapping is added. Auto panning needs to be implemented on idle.